backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-jose
passlib[bcrypt]
mysql-connector-python
python-dotenv
email-validator
flask
PyJWT
email-validator
python-multipart
